+++ src/dynamics/ccd/ccd_solver.rs
@@ -4,7 +4,7 @@ use crate::geometry::{ColliderParent, ColliderSet, CollisionEvent, NarrowPhase};
 use crate::math::Real;
 use crate::parry::utils::SortedPair;
 use crate::pipeline::{EventHandler, QueryPipeline};
-use crate::prelude::{updaters, ActiveEvents, CollisionEventFlags};
+use crate::prelude::{query_pipeline_generators, ActiveEvents, CollisionEventFlags};
 use parry::query::{DefaultQueryDispatcher, QueryDispatcher};
 use parry::utils::hashmap::HashMap;
 use std::collections::BinaryHeap;
@@ -117,8 +117,13 @@ impl CCDSolver {
         narrow_phase: &NarrowPhase,
     ) -> Option<Real> {
         // Update the query pipeline.
-        self.query_pipeline
-            .update_with_mode(updaters::SweepTestWithNextPosition { bodies, colliders });
+        self.query_pipeline.update_with_generator(
+            query_pipeline_generators::SweptAabbWithPredictedPosition {
+                bodies,
+                colliders,
+                dt,
+            },
+        );
 
         let mut pairs_seen = HashMap::default();
         let mut min_toi = dt;
@@ -235,8 +240,9 @@ impl CCDSolver {
         let mut min_overstep = dt;
 
         // Update the query pipeline.
-        self.query_pipeline
-            .update_with_mode(updaters::SweepTestWithNextPosition { bodies, colliders });
+        self.query_pipeline.update_with_generator(
+            query_pipeline_generators::SweptAabbWithNextPosition { bodies, colliders },
+        );
 
         /*
          *